refactor: dtos.ts -> types.d.ts 네임 변경
